fix #85; do not clean the file up after each archive job
