+++ refgenieserver/main.py
@@ -1,5 +1,5 @@
 from .const import *
-from .helpers import build_parser
+from .helpers import build_parser, purge_nonservable
 from .server_builder import archive
 from refgenconf import RefGenConf, select_genome_config
 from fastapi import FastAPI
@@ -41,6 +41,8 @@ def main():
         archive(rgc, arp, args.force, args.remove, selected_cfg, args.genomes_desc)
     elif args.command == "serve":
         # the router imports need to be after the RefGenConf object is declared
+        with rgc as r:
+            purge_nonservable(r)
         from .routers import version1, version2
         app.include_router(version1.router)
         app.include_router(version1.router, prefix="/v1")
